P4.3: 共享 UI 上真屏——6/6 帧 CRC 固件=模拟器=golden 逐位全等；修 LVGL I1 偏移/reshape 两 bug；presenter 格式修补

+++ shared/presenter/cdt_presenter.c
@@ -165,8 +165,8 @@ static void fmt_duration(char *dst, size_t dstsz, uint64_t ms)
     uint32_t m = (uint32_t)((total_s % 3600u) / 60u);
     uint32_t s = (uint32_t)(total_s % 60u);
 
-    if (h > 0) snprintf(dst, dstsz, "%02u:%02u:%02u", (unsigned)h, m, s);
-    else snprintf(dst, dstsz, "%02u:%02u", m, s);
+    if (h > 0) snprintf(dst, dstsz, "%02u:%02u:%02u", (unsigned)h, (unsigned)m, (unsigned)s);
+    else snprintf(dst, dstsz, "%02u:%02u", (unsigned)m, (unsigned)s);
 }
 
 static const char *status_label_of(cdt_thread_state_t st)
